Change piID to a C.int type

+++ rpi/i2c.go
@@ -32,7 +32,7 @@ func (s *piPigpioI2CHandle) Write(ctx context.Context, tx []byte) error {
 	txPtr := C.CBytes(tx)
 	defer C.free(txPtr)
 
-	ret := int(C.i2c_write_device(C.int(s.bus.pi.piID), s.handle, (*C.char)(txPtr), (C.uint)(len(tx))))
+	ret := int(C.i2c_write_device(s.bus.pi.piID, s.handle, (*C.char)(txPtr), (C.uint)(len(tx))))
 
 	if ret != 0 {
 		return picommon.ConvertErrorCodeToMessage(ret, "error with i2c write")
@@ -47,7 +47,7 @@ func (s *piPigpioI2CHandle) Read(ctx context.Context, count int) ([]byte, error)
 	rxPtr := C.CBytes(rx)
 	defer C.free(rxPtr)
 
-	ret := int(C.i2c_read_device(C.int(s.bus.pi.piID), s.handle, (*C.char)(rxPtr), (C.uint)(count)))
+	ret := int(C.i2c_read_device(s.bus.pi.piID, s.handle, (*C.char)(rxPtr), (C.uint)(count)))
 
 	if ret <= 0 {
 		return nil, picommon.ConvertErrorCodeToMessage(ret, "error with i2c read")
@@ -57,7 +57,7 @@ func (s *piPigpioI2CHandle) Read(ctx context.Context, count int) ([]byte, error)
 }
 
 func (s *piPigpioI2CHandle) ReadByteData(ctx context.Context, register byte) (byte, error) {
-	res := C.i2c_read_byte_data(C.int(s.bus.pi.piID), s.handle, C.uint(register))
+	res := C.i2c_read_byte_data(s.bus.pi.piID, s.handle, C.uint(register))
 	if res < 0 {
 		return 0, picommon.ConvertErrorCodeToMessage(int(res), "error in ReadByteData")
 	}
@@ -65,7 +65,7 @@ func (s *piPigpioI2CHandle) ReadByteData(ctx context.Context, register byte) (by
 }
 
 func (s *piPigpioI2CHandle) WriteByteData(ctx context.Context, register, data byte) error {
-	res := C.i2c_write_byte_data(C.int(s.bus.pi.piID), s.handle, C.uint(register), C.uint(data))
+	res := C.i2c_write_byte_data(s.bus.pi.piID, s.handle, C.uint(register), C.uint(data))
 	if res != 0 {
 		return picommon.ConvertErrorCodeToMessage(int(res), "error in WriteByteData")
 	}
@@ -79,7 +79,7 @@ func (s *piPigpioI2CHandle) ReadBlockData(ctx context.Context, register byte, nu
 
 	data := make([]byte, numBytes)
 	response := C.i2c_read_i2c_block_data(
-		C.int(s.bus.pi.piID), s.handle, C.uint(register), (*C.char)(&data[0]), C.uint(numBytes))
+		s.bus.pi.piID, s.handle, C.uint(register), (*C.char)(&data[0]), C.uint(numBytes))
 	if response < 0 {
 		return nil, picommon.ConvertErrorCodeToMessage(int(response), "error in ReadBlockData")
 	}
@@ -93,7 +93,7 @@ func (s *piPigpioI2CHandle) WriteBlockData(ctx context.Context, register byte, d
 	}
 
 	response := C.i2c_write_i2c_block_data(
-		C.int(s.bus.pi.piID), s.handle, C.uint(register), (*C.char)(&data[0]), C.uint(numBytes))
+		s.bus.pi.piID, s.handle, C.uint(register), (*C.char)(&data[0]), C.uint(numBytes))
 	if response != 0 {
 		return picommon.ConvertErrorCodeToMessage(int(response), "error in WriteBlockData")
 	}
@@ -118,6 +118,6 @@ func (s *piPigpioI2C) OpenHandle(addr byte) (buses.I2CHandle, error) {
 }
 
 func (s *piPigpioI2CHandle) Close() error {
-	C.i2c_close(C.int(s.bus.pi.piID), s.handle)
+	C.i2c_close(s.bus.pi.piID, s.handle)
 	return nil
 }
